working process

+++ app/src/main/java/com/mincor/kviper/models/Wind.kt
@@ -2,5 +2,5 @@ package com.mincor.kviper.models
 
 data class Wind (
     var speed: Double? = null,
-    var deg: Int? = null
+    var deg: Double? = null
 )
